pyproject.toml
[project]
name = "paperpilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "ollama>=0.6.1",
    "pydantic-settings>=2.12.0",
    "rich>=14.3.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
